:recycle: api: move controllers and routes into enrolment context folder
