+++ api/tests/certification/enrolment/unit/application/attendance-sheet-controller_test.js
@@ -1,4 +1,4 @@
-import { attendanceSheetController } from '../../../../../src/certification/session/application/attendance-sheet-controller.js';
+import { attendanceSheetController } from '../../../../../src/certification/enrolment/application/attendance-sheet-controller.js';
 import { usecases } from '../../../../../src/certification/session/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../../test-helper.js';
 import { getI18n } from '../../../../tooling/i18n/i18n.js';
