+++ api/tests/certification/enrolment/unit/application/attendance-sheet-route_test.js
@@ -1,6 +1,6 @@
 import { authorization } from '../../../../../lib/application/preHandlers/authorization.js';
-import { attendanceSheetController } from '../../../../../src/certification/session/application/attendance-sheet-controller.js';
-import * as moduleUnderTest from '../../../../../src/certification/session/application/attendance-sheet-route.js';
+import { attendanceSheetController } from '../../../../../src/certification/enrolment/application/attendance-sheet-controller.js';
+import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/attendance-sheet-route.js';
 import { NotFoundError } from '../../../../../src/shared/domain/errors.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
