+++ api/tests/certification/enrolment/unit/application/certification-candidate-route_test.js
@@ -1,7 +1,7 @@
 import { authorization } from '../../../../../lib/application/preHandlers/authorization.js';
 import { NotFoundError } from '../../../../../lib/domain/errors.js';
-import { certificationCandidateController } from '../../../../../src/certification/session/application/certification-candidate-controller.js';
-import * as moduleUnderTest from '../../../../../src/certification/session/application/certification-candidate-route.js';
+import { certificationCandidateController } from '../../../../../src/certification/enrolment/application/certification-candidate-controller.js';
+import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/certification-candidate-route.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
 describe('Unit | Application | Sessions | Routes', function () {
